+++ internal/backend/remote-state/azure/meta-test/variables.tf
@@ -27,3 +27,9 @@ variable "ssh_pub_key_path" {
   type        = string
   description = "The file path on the local file system where this user's public SSH key is located. This is for ssh-ing into the VM used for MSI testing, and so it is only relevant if use_msi is set to true."
 }
+
+variable "use_ado" {
+  default     = false
+  type        = bool
+  description = "Set this to setup a pre-existing Azure DevOps org for integration testing. "
+}