+++ internal/command/views/state_test.go
@@ -12,7 +12,9 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/opentofu/opentofu/internal/addrs"
 	"github.com/opentofu/opentofu/internal/command/arguments"
+	"github.com/opentofu/opentofu/internal/states"
 	"github.com/opentofu/opentofu/internal/tfdiags"
+	regaddr "github.com/opentofu/registry-address/v2"
 )
 
 func TestStateViews(t *testing.T) {
@@ -165,6 +167,166 @@ func TestStateViews(t *testing.T) {
 				},
 			},
 			wantStdout: withNewline("Successfully moved 1 object(s)."),
+		},
+		"printPulledState": {
+			viewCall: func(state State) {
+				state.PrintPulledState(`{"version":4,"terraform_version":"1.11.5","serial":9,"lineage":"9ba8c556-ae6c-20ee-f6ed-b57c7cc04dcd","outputs":{},"resources":[]}`)
+			},
+			wantJson: []map[string]any{
+				{
+					"@level":   "error",
+					"@message": "printing the pulled state is not available in the JSON view. The `tofu state pull` should not be configured with the `-json` flag",
+					"@module":  "tofu.ui",
+				},
+			},
+			wantStdout: withNewline(`{"version":4,"terraform_version":"1.11.5","serial":9,"lineage":"9ba8c556-ae6c-20ee-f6ed-b57c7cc04dcd","outputs":{},"resources":[]}`),
+		},
+		"noMatchingResourcesForProviderReplacement": {
+			viewCall: func(state State) {
+				state.NoMatchingResourcesForProviderReplacement()
+			},
+			wantJson: []map[string]any{
+				{
+					"@level":   "info",
+					"@message": "No matching resources found",
+					"@module":  "tofu.ui",
+				},
+			},
+			wantStdout: withNewline(`No matching resources found.`),
+		},
+		"replaceProviderOverview": {
+			viewCall: func(state State) {
+				state.ReplaceProviderOverview(
+					regaddr.NewProvider("registry1.org", "ns", "prov"),
+					regaddr.NewProvider("registry2.org", "ns", "prov"),
+					[]*states.Resource{
+						{
+							Addr: addrs.AbsResource{Resource: addrs.Resource{
+								Mode: addrs.ManagedResourceMode,
+								Type: "res",
+								Name: "foo",
+							}},
+						},
+					},
+				)
+			},
+			wantJson: []map[string]any{
+				{
+					"@level":    "info",
+					"@message":  "OpenTofu will replace provider from registry1.org/ns/prov to registry2.org/ns/prov for 1 resources",
+					"resources": []any{"res.foo"},
+					"@module":   "tofu.ui",
+					"type":      "replace_provider",
+					"from":      "registry1.org/ns/prov",
+					"to":        "registry2.org/ns/prov",
+				},
+			},
+			wantStdout: `OpenTofu will perform the following actions:
+
+  ~ Updating provider:
+    - registry1.org/ns/prov
+    + registry2.org/ns/prov
+
+Changing 1 resources:
+
+  res.foo
+`,
+		},
+		"replaceProviderCancelled": {
+			viewCall: func(state State) {
+				state.ReplaceProviderCancelled()
+			},
+			wantJson: []map[string]any{
+				{
+					"@level":   "info",
+					"@message": "Cancelled replacing providers",
+					"@module":  "tofu.ui",
+				},
+			},
+			wantStdout: withNewline(`Cancelled replacing providers.`),
+		},
+		"providerReplaced": {
+			viewCall: func(state State) {
+				state.ProviderReplaced(2)
+			},
+			wantJson: []map[string]any{
+				{
+					"@level":   "info",
+					"@message": "Successfully replaced provider for 2 resources",
+					"@module":  "tofu.ui",
+				},
+			},
+			wantStdout: withNewline(`Successfully replaced provider for 2 resources.`),
+		},
+		"resourceRemoveStatus with dryRun=true": {
+			viewCall: func(state State) {
+				state.ResourceRemoveStatus(true, "test_res.name1")
+			},
+			wantJson: []map[string]any{
+				{
+					"@level":   "info",
+					"@message": "Would remove test_res.name1",
+					"@module":  "tofu.ui",
+				},
+			},
+			wantStdout: withNewline("Would remove test_res.name1"),
+		},
+		"resourceRemoveStatus with dryRun=false": {
+			viewCall: func(state State) {
+				state.ResourceRemoveStatus(false, "test_res.name1")
+			},
+			wantJson: []map[string]any{
+				{
+					"@level":   "info",
+					"@message": "Removed test_res.name1",
+					"@module":  "tofu.ui",
+				},
+			},
+			wantStdout: withNewline("Removed test_res.name1"),
+		},
+		"dryRunRemovedStatus with 0 resources": {
+			viewCall: func(state State) {
+				state.DryRunRemovedStatus(0)
+			},
+			wantJson: []map[string]any{
+				{
+					"@level":   "info",
+					"@message": "Would have removed nothing",
+					"@module":  "tofu.ui",
+				},
+			},
+			wantStdout: withNewline("Would have removed nothing."),
+		},
+		"dryRunRemovedStatus with >0 resources": {
+			viewCall: func(state State) {
+				state.DryRunRemovedStatus(1)
+			},
+			wantJson: []map[string]any{
+				{},
+			},
+			wantStdout: "",
+		},
+		"removeFinalStatus with 0 resources": {
+			viewCall: func(state State) {
+				state.RemoveFinalStatus(0)
+			},
+			wantJson: []map[string]any{
+				{},
+			},
+			wantStdout: "",
+		},
+		"removeFinalStatus with >0 resources": {
+			viewCall: func(state State) {
+				state.RemoveFinalStatus(2)
+			},
+			wantJson: []map[string]any{
+				{
+					"@level":   "info",
+					"@message": "Successfully removed 2 resource instance(s)",
+					"@module":  "tofu.ui",
+				},
+			},
+			wantStdout: withNewline("Successfully removed 2 resource instance(s)."),
 		},
 		// Diagnostics
 		"warning": {